Rename model to variables to follow standard practice
