+++ openfisca_tunisia/variables/caracteristiques_socio_demographiques/demographie.py
@@ -1,7 +1,7 @@
 from numpy import datetime64
 
 
-from openfisca_tunisia.model.base import *
+from openfisca_tunisia.variables.base import *
 
 
 class age(Variable):
