+++ openfisca_tunisia/variables/prelevements_obligatoires/cotisations_sociales.py
@@ -2,7 +2,7 @@ from __future__ import division
 
 from numpy import zeros
 
-from openfisca_tunisia.model.base import *  # noqa analysis:ignore
+from openfisca_tunisia.variables.base import *  # noqa analysis:ignore
 
 
 class TypesRegimeSecuriteSociale(Enum):
