+++ openfisca_tunisia/variables/revenus/capital/foncier.py
